Feature_extraction/requirements.txt
torch
biopython
numpy
pandas
scikit-learn
